+++ orig/trunk/vnc_winsrc/wsconfig/Settings.cpp
@@ -31,7 +31,7 @@ bool Settings::saveToStorage(SettingsManager *sm)
   bool saveResult = true;
   // Save Port Mapping Vector
   tstring strPortMapping = _T("");
-  for (int i = 0; i < m_vPortMapping.size(); i++) {
+  for (size_t i = 0; i < m_vPortMapping.size(); i++) {
     strPortMapping += m_vPortMapping.at(i).toString();
     if (i != m_vPortMapping.size() - 1)
       strPortMapping += _T(",");
